requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjwt==2.8.0
slowapi==0.1.9
aiosqlite==0.19.0
bcrypt==4.0.1